improve filter and sorting functionality
