+++ src/components/Layout.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function Layout({ children }) {
+  return (
+    <div>
+      {/* Add your header, footer, or other layout components here */}
+      {children}
+    </div>
+  );
+}
+
+export default Layout;
